Extracting get_block_at from service to client
