+++ crates/adapters/celestia/src/types/mod.rs
@@ -245,12 +245,27 @@ impl FilteredCelestiaBlock {
         rollup_batch_data: NamespaceRelevantData,
         rollup_proof_data: NamespaceRelevantData,
         header: celestia_types::ExtendedHeader,
+        // TODO: Why it is Result?
     ) -> anyhow::Result<Self> {
         Ok(FilteredCelestiaBlock {
             header: CelestiaHeader::new(header.dah, header.header.into()),
             rollup_batch_data,
             rollup_proof_data,
         })
+    }
+
+    // TODO: Should we return extended header from Twinkle?
+    #[cfg(feature = "native")]
+    pub(crate) fn new_with_short_header(
+        rollup_batch_data: NamespaceRelevantData,
+        rollup_proof_data: NamespaceRelevantData,
+        header: CelestiaHeader,
+    ) -> Self {
+        FilteredCelestiaBlock {
+            header,
+            rollup_batch_data,
+            rollup_proof_data,
+        }
     }
 }
